feat: organized award components & remove blob background + migration of animations to lib & global import cleanup
