+++ pages/bulletin/[slug].js
@@ -12,7 +12,7 @@ import Head from 'next/head';
 import Image from 'next/image';
 import Link from 'next/link';
 import { PortableText } from '@portabletext/react';
-import { _Transition_Page } from '../../components/_Animations';
+import { _Transition_Page } from '../../lib/animations';
 import { client } from '../../lib/sanity';
 import dayjs from 'dayjs';
 import urlBuilder from '@sanity/image-url';
