+++ pages/gallery/[slug].js
@@ -3,7 +3,7 @@ import { CgChevronLeft } from 'react-icons/cg';
 import Head from 'next/head';
 import Image from 'next/image';
 import Link from 'next/link';
-import { _Transition_Page } from '../../components/_Animations';
+import { _Transition_Page } from '../../lib/animations';
 import { client } from '../../lib/sanity';
 import dayjs from 'dayjs';
 import { motion } from 'framer-motion';
