+++ pages/thesis/[slug].js
@@ -7,7 +7,7 @@ import Image from 'next/image';
 import Link from 'next/link';
 import Script from 'next/script';
 import { PortableText } from '@portabletext/react';
-import { _Transition_Page } from '../../components/_Animations';
+import { _Transition_Page } from '../../lib/animations';
 import { client } from '../../lib/sanity';
 import dayjs from 'dayjs';
 import urlBuilder from '@sanity/image-url';
