+++ pages/thesis/index.js
@@ -5,7 +5,7 @@ import DeptSection from '../../components/Thesis/DeptSection';
 import Head from '../../components/Head';
 import ThesisCard from '../../components/Card/Thesis';
 import TopGradient from '../../components/TopGradient';
-import { _Transition_Page } from '../../components/_Animations';
+import { _Transition_Page } from '../../lib/animations';
 import { motion, AnimatePresence } from 'framer-motion';
 import { usePrefetcher } from '../../components/Prefetcher';
 import { CgArrowUp, CgArrowDown, CgSearch, CgClose } from 'react-icons/cg';
